refactor: изменена структура папок
