+++ src/app/app.routes.ts
@@ -13,7 +13,8 @@ const routes: Routes = [
   },
   {
     path: 'login',
-    loadChildren: () => import('./auth/auth.module').then((m) => m.AuthModule),
+    loadChildren: () =>
+      import('./pages/auth/auth.module').then((m) => m.AuthModule),
     canActivate: [notAuthrorizedGuard],
   },
 ];
