+++ src/app/pages/auth/components/password-input/password-input.component.scss
