+++ src/app/pages/auth/pipes/format-password-requirement.pipe.ts
