+++ src/app/pages/auth/services/auth/auth.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { AuthToken } from '../../models/auth-token.model';
-import ENDPOINTS from '../../../shared/enums/endpoints';
+import ENDPOINTS from '../../../../shared/enums/endpoints';
 
 @Injectable({
   providedIn: 'root',
