+++ src/app/pages/auth/validators/password.validator.ts
